+++ odoo/local-src/connector_esb/tests/test_export_stock_update.py
@@ -2,6 +2,10 @@
 # Copyright 2018 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
+import mock
+import os
+import requests
+
 from datetime import datetime, timedelta
 from odoo.tests.common import SavepointCase
 
@@ -10,6 +14,9 @@ class ExportStockUpdateTestCase(SavepointCase):
 
     def setUp(self):
         super(ExportStockUpdateTestCase, self).setUp()
+        os.environ['ODOO_ESB_WS_USER'] = 'ws_user'
+        os.environ['ODOO_ESB_WS_BASE_URL'] = 'https://test.com'
+        os.environ['ODOO_ESB_WS_PWD'] = 'pwd'
         self.backend_model = self.env['esb.backend']
         self.backend = self.backend_model.get_singleton()
         self.setup_records()
@@ -200,3 +207,18 @@ class ExportStockUpdateTestCase(SavepointCase):
             exporter = work.component(usage='record.exporter.cron')
             items = exporter.get_items(None)
         self.assertEqual(len(items), 2)
+
+    def post_ret_status(url, data, headers, auth):
+        resp = requests.Response()
+        resp.status_code = 200
+        resp.json = lambda: '{"status" : "OK", “code” : “200”, "items": []}'
+        return resp
+
+    @mock.patch('requests.post', side_effect=post_ret_status)
+    def test_record_exporter(self, post):
+        """Test export of a sale order catching the post request."""
+        with self.backend.work_on(self.model._name,
+                                  timestamp=self.timestamp) as work:
+            exporter = work.component(usage='record.exporter.cron')
+            exporter.run()
+        post.assert_called_once()